django_news/requirements.txt
Django==4.0.3
Pillow
django-ckeditor